Move getPropertyFromConnectionString to utils
