+++ src/tree/AttachedStorageAccountTreeItem.ts
@@ -11,11 +11,11 @@ import * as path from 'path';
 import { Uri } from 'vscode';
 import { AzureParentTreeItem, AzureTreeItem } from 'vscode-azureextensionui';
 import { emulatorAccountName, emulatorConnectionString, emulatorKey, getResourcesPath } from '../constants';
+import { getPropertyFromConnectionString } from '../utils/getPropertyFromConnectionString';
 import { localize } from '../utils/localize';
 import { AttachedAccountRoot } from './AttachedStorageAccountsTreeItem';
 import { BlobContainerGroupTreeItem } from './blob/BlobContainerGroupTreeItem';
 import { FileShareGroupTreeItem } from './fileShare/FileShareGroupTreeItem';
-import { getPropertyFromConnectionString } from './getPropertyFromConnectionString';
 import { IStorageRoot } from './IStorageRoot';
 import { QueueGroupTreeItem } from './queue/QueueGroupTreeItem';
 import { StorageAccountTreeItem, WebsiteHostingStatus } from './StorageAccountTreeItem';
