+++ src/utils/getPropertyFromConnectionString.ts
